+++ code/test/sixsq/nuvla/auth/acl_resource_test.clj
@@ -49,6 +49,7 @@
                                ::a/view-data
                                ::a/query
                                ::a/bulk-delete
+                               ::a/bulk-action
                                ::a/add} {:claims #{"user/user1"}}
 
                              #{::a/edit-acl
@@ -61,6 +62,7 @@
                                ::a/view-data
                                ::a/query
                                ::a/bulk-delete
+                               ::a/bulk-action
                                ::a/add} {:claims #{"group/nuvla-admin"}}
 
                              #{::a/view-acl
@@ -82,6 +84,7 @@
                                ::a/view-data
                                ::a/query
                                ::a/bulk-delete
+                               ::a/bulk-action
                                ::a/add} {:claims #{"user/user1", "group/role2"}}
 
                              #{::a/delete
@@ -145,7 +148,12 @@
       (is (isa? a/rights-hierarchy right right)))
 
     ;; ::edit-acl covers everything except ::query and ::add
-    (doseq [right (disj all-rights ::a/edit-acl ::a/query ::a/add ::a/bulk-delete)]
+    (doseq [right (disj all-rights
+                        ::a/edit-acl
+                        ::a/query
+                        ::a/add
+                        ::a/bulk-delete
+                        ::a/bulk-action)]
       (is (isa? a/rights-hierarchy ::a/edit-acl right))
       (is (not (isa? a/rights-hierarchy right ::a/edit-acl)))))
 
@@ -153,7 +161,8 @@
                               ::a/edit-acl :edit-acl
                               ::a/query :query
                               ::a/add :add
-                              ::a/bulk-delete :bulk-delete))]
+                              ::a/bulk-delete :bulk-delete
+                              ::a/bulk-action :bulk-action))]
     (is (isa? a/rights-hierarchy ::a/edit-acl right))
     (is (not (isa? a/rights-hierarchy right ::a/edit-acl)))))