+++ code/test/sixsq/nuvla/server/resources/nuvlabox_1_lifecycle_test.clj
@@ -91,7 +91,7 @@
                             session
                             (content-type "application/json"))
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")]
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")]
 
       (let [nuvlabox-id  (-> session-owner
                              (request base-uri
@@ -140,10 +140,10 @@
     (let [session       (-> (ltu/ring-app)
                             session
                             (content-type "application/json"))
-          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
-          session-anon  (header session authn-info-header "unknown group/nuvla-anon")]
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")]
 
       (doseq [session [session-admin session-owner]]
         (let [nuvlabox-id  (-> session
@@ -305,10 +305,10 @@
     (let [session       (-> (ltu/ring-app)
                             session
                             (content-type "application/json"))
-          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
-          session-anon  (header session authn-info-header "unknown group/nuvla-anon")
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")
           tags          #{"tag-1", "tag-2"}]
 
       (doseq [session [session-admin session-owner]]
@@ -624,10 +624,10 @@
                             session
                             (content-type "application/json"))
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
-          session-anon  (header session authn-info-header "unknown group/nuvla-anon")
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")
           user-beta     "user/beta"
-          session-beta  (header session authn-info-header (str user-beta " group/nuvla-user group/nuvla-anon"))]
+          session-beta  (header session authn-info-header (str user-beta " " user-beta " group/nuvla-user group/nuvla-anon"))]
 
       (let [nuvlabox-id  (-> session-owner
                              (request base-uri
@@ -831,8 +831,8 @@
                             session
                             (content-type "application/json"))
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
-          session-anon  (header session authn-info-header "unknown group/nuvla-anon")]
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "unknown unknown group/nuvla-anon")]
 
       (let [nuvlabox-id      (-> session-owner
                                  (request base-uri
@@ -1059,10 +1059,10 @@
     (let [session       (-> (ltu/ring-app)
                             session
                             (content-type "application/json"))
-          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
-          session-anon  (header session authn-info-header "unknown group/nuvla-anon")]
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")]
 
       (let [infra-srvc-vpn-create {:template {:href      (str infra-service-tpl/resource-type "/"
                                                               infra-srvc-tpl-vpn/method)
@@ -1121,7 +1121,7 @@
             (ltu/is-status 200))
 
         (let [session-nuvlabox  (header session authn-info-header
-                                        (str nuvlabox-id
+                                        (str nuvlabox-id " " nuvlabox-id
                                              " group/nuvla-nuvlabox group/nuvla-anon"))
               commission        (-> session-owner
                                     (request nuvlabox-url)
@@ -1202,10 +1202,10 @@
     (let [session       (-> (ltu/ring-app)
                             session
                             (content-type "application/json"))
-          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
-          session-anon  (header session authn-info-header "unknown group/nuvla-anon")]
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")]
 
       (let [infra-srvc-vpn-create {:template {:href      (str infra-service-tpl/resource-type "/"
                                                               infra-srvc-tpl-vpn/method)
@@ -1264,7 +1264,7 @@
             (ltu/is-status 200))
 
         (let [session-nuvlabox  (header session authn-info-header
-                                        (str nuvlabox-id
+                                        (str nuvlabox-id " " nuvlabox-id
                                              " group/nuvla-nuvlabox group/nuvla-anon"))
               commission        (-> session-owner
                                     (request nuvlabox-url)
@@ -1343,8 +1343,8 @@
                                 session
                                 (content-type "application/json"))
 
-          session-admin     (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-          session-owner     (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
+          session-admin     (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-owner     (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
 
           customer-base-uri (str p/service-context customer/resource-type)]
 
@@ -1406,8 +1406,8 @@
                             session
                             (content-type "application/json"))
 
-          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
 
           nuvlabox-id   (-> session-owner
                             (request base-uri
@@ -1467,10 +1467,10 @@
     (let [session       (-> (ltu/ring-app)
                             session
                             (content-type "application/json"))
-          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
 
-          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
-          session-anon  (header session authn-info-header "unknown group/nuvla-anon")]
+          session-owner (header session authn-info-header "user/alpha user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")]
 
       (let [nuvlabox-id  (-> session-owner
                              (request base-uri
@@ -1496,7 +1496,7 @@
             (ltu/is-status 200))
 
         (let [session-nuvlabox (header session authn-info-header
-                                       (str nuvlabox-id
+                                       (str nuvlabox-id " " nuvlabox-id
                                             " group/nuvla-nuvlabox group/nuvla-anon"))
               commission       (-> session-owner
                                    (request nuvlabox-url)