+++ code/test/sixsq/nuvla/server/resources/session_password_lifecycle_test.clj
@@ -78,9 +78,9 @@
 
   (let [app              (ltu/ring-app)
         session-json     (content-type (session app) "application/json")
-        session-anon     (header session-json authn-info-header "user/unknown group/nuvla-anon")
+        session-anon     (header session-json authn-info-header "user/unknown user/unknown group/nuvla-anon")
         session-user     (header session-json authn-info-header "user group/nuvla-user")
-        session-admin    (header session-json authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-admin    (header session-json authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
 
         href             (str st/resource-type "/password")
 
@@ -287,8 +287,8 @@
 
   (let [app                (ltu/ring-app)
         session-json       (content-type (session app) "application/json")
-        session-anon       (header session-json authn-info-header "user/unknown group/nuvla-anon")
-        session-admin      (header session-json authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-anon       (header session-json authn-info-header "user/unknown user/unknown group/nuvla-anon")
+        session-admin      (header session-json authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
 
         href               (str st/resource-type "/password")