+++ code/test/sixsq/nuvla/server/resources/user_email_password_lifecycle_test.clj
@@ -40,9 +40,9 @@
         session         (-> (ltu/ring-app)
                             session
                             (content-type "application/json"))
-        session-admin   (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-user    (header session authn-info-header "user/jane group/nuvla-user group/nuvla-anon")
-        session-anon    (header session authn-info-header "user/unknown group/nuvla-anon")]
+        session-admin   (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-user    (header session authn-info-header "user/jane user/jane group/nuvla-user group/nuvla-anon")
+        session-anon    (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")]
 
     (with-redefs [email-utils/extract-smtp-cfg
                                       (fn [_] {:host "smtp@example.com"
@@ -134,7 +134,7 @@
                                        (ltu/body->edn)
                                        (ltu/is-status 201))
               user-id              (ltu/body-resource-id resp)
-              session-created-user (header session authn-info-header (str user-id " group/nuvla-user group/nuvla-anon"))
+              session-created-user (header session authn-info-header (str user-id " " user-id " group/nuvla-user group/nuvla-anon"))
 
               {credential-id :credential-password,
                email-id      :email :as user} (-> session-created-user
@@ -270,8 +270,8 @@
       (let [session            (-> (ltu/ring-app)
                                   session
                                   (content-type "application/json"))
-           session-admin      (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-           session-anon       (header session authn-info-header "user/unknown group/nuvla-anon")
+           session-admin      (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+           session-anon       (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")
 
            href               (str user-tpl/resource-type "/" email-password/registration-method)
 
@@ -308,7 +308,7 @@
                               (ltu/body->edn)
                               (ltu/is-status 201))
              user-id      (ltu/body-resource-id resp)
-             session-user (header session authn-info-header (str user-id " group/nuvla-user group/nuvla-anon"))]
+             session-user (header session authn-info-header (str user-id " " user-id " group/nuvla-user group/nuvla-anon"))]
 
          ; credential password is created and visible by the created user
          (-> session-user