+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+### Added
+
+  - Deployment - Support bulk update
+  - Bulk - Bulk operation support
+
+### Changed
+
+  - Header authentication extended to support active-claim
+
 ## [5.11.1] - 2021-03-22
 
 ### Changed